+++ dungeonCrawler/ECS/Projectile/Effect/ProjectileHitEffect.swift
@@ -1,0 +1,13 @@
+//
+//  ProjectileHitEffect.swift
+//  dungeonCrawler
+//
+//  Created by Letian on 12/4/26.
+//
+
+import Foundation
+import simd
+
+public protocol ProjectileHitEffect {
+    func apply(context: HitContext)
+}
