+++ dungeonCrawler/ECS/Projectile/Effect/ZoneBase.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public struct ZoneBase {
+    let textureName: String
+    let radius: Float
+    let damagePerSecond: Float
+    let duration: Float
+}
