+++ dungeonCrawler/ECS/Projectile/ProjectileSystem.swift
@@ -1,10 +1,3 @@
-//
-//  ProjectileSystem.swift
-//  dungeonCrawler
-//
-//  Created by Letian on 20/3/26.
-//
-
 import Foundation
 import simd
