--- dungeonCrawler/ECS/Projectile/ZoneContext.swift
@@ -1,14 +1,0 @@
-////
-////  ZoneContext.swift
-////  dungeonCrawler
-////
-////  Created by Letian on 12/4/26.
-////
-//
-//import Foundation
-//
-//public struct ZoneContext: HitContext {
-//    var center: SIMD2<Float>
-//    let world: World
-//    let zoneBase: ZoneBase
-//}
